# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/it/security/policy": {
      "get": {
        "operationId": "SecurityController_getPolicy",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "سیاست رمز عبور و امنیت فعلی",
        "tags": [
          "it-manager"
        ]
      },
      "patch": {
        "operationId": "SecurityController_updatePolicy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecurityPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "به‌روزرسانی سیاست رمز عبور و امنیت",
        "tags": [
          "it-manager"
        ]
      }
    },
...
    "/it/security/sessions/logout-all": {
      "post": {
        "operationId": "SecurityController_logoutAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutAllDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "خروج اجباری همه نشست‌های فعال سایت",
        "tags": [
          "it-manager"
        ]
      }
    },
...
    "/search/flights": {
      "get": {
        "operationId": "SearchController_flights",
        "parameters": [
          {
            "name": "origin",
            "required": true,
            "in": "query",
            "schema": {
              "example": "THR",
              "type": "string"
            }
          },
          {
            "name": "dest",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MHD",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جستجوی پرواز بین دو فرودگاه در یک روز مشخص",
        "tags": [
          "search"
        ]
      }
    },
    "/search/flights/{id}/seatmap": {
      "get": {
        "operationId": "SearchController_seatMap",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "نقشه صندلی برای انتخاب صندلی هنگام خرید",
        "tags": [
          "search"
        ]
      }
    },
    "/search/price-calendar": {
      "get": {
        "operationId": "SearchController_priceCalendar",
        "parameters": [
          {
            "name": "origin",
            "required": true,
            "in": "query",
            "schema": {
              "example": "THR",
              "type": "string"
            }
          },
          {
            "name": "dest",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MHD",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "تقویم قیمت ۷روزه برای نوار نتایج پرواز",
        "tags": [
          "search"
        ]
      }
    },
    "/search/advisory": {
      "get": {
        "operationId": "SearchController_getAdvisory",
        "parameters": [
          {
            "name": "origin",
            "required": true,
            "in": "query",
            "schema": {
              "example": "THR",
              "type": "string"
            }
          },
          {
            "name": "dest",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MHD",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "رادار هوشمند قیمت — توصیه خرید/انتظار (advisory، degrade امن)",
        "tags": [
          "search"
        ]
      }
    },
...
    "/blog/admin/posts": {
      "get": {
        "operationId": "BlogAdminController_list",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEWS",
                "GUIDE",
                "DEST",
                "OFFERS",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست همهٔ مقالات (با فیلتر دسته‌بندی)",
        "tags": [
          "blog"
        ]
      },
      "post": {
        "operationId": "BlogAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "ایجاد مقالهٔ جدید",
        "tags": [
          "blog"
        ]
      }
    },
    "/blog/admin/posts/{id}": {
      "get": {
        "operationId": "BlogAdminController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جزئیات یک مقاله برای ویرایش",
        "tags": [
          "blog"
        ]
      },
      "patch": {
        "operationId": "BlogAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ویرایش مقاله",
        "tags": [
          "blog"
        ]
      },
      "delete": {
        "operationId": "BlogAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "حذف نرم مقاله",
        "tags": [
          "blog"
        ]
      }
    },
    "/blog/posts": {
      "get": {
        "operationId": "BlogPublicController_list",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEWS",
                "GUIDE",
                "DEST",
                "OFFERS"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "فهرست مقالات منتشرشده",
        "tags": [
          "blog"
        ]
      }
    },
    "/blog/posts/{slug}": {
      "get": {
        "operationId": "BlogPublicController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "جزئیات یک مقالهٔ منتشرشده",
        "tags": [
          "blog"
        ]
      }
    },
    "/blog/covers/{fileId}": {
      "get": {
        "operationId": "BlogPublicController_cover",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "دریافت تصویر کاور مقالهٔ منتشرشده",
        "tags": [
          "blog"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateSecurityPolicyDto": {
        "type": "object",
        "properties": {
          "minLength": {
            "type": "number",
            "example": 10
          },
          "expiryDays": {
            "type": "number",
            "example": 90
          },
          "maxAttempts": {
            "type": "number",
            "example": 5
          },
          "requireUppercase": {
            "type": "boolean"
          },
          "requireNumber": {
            "type": "boolean"
          },
          "requireSymbol": {
            "type": "boolean"
          },
          "blockReuse": {
            "type": "boolean"
          },
          "staffTwoFactorMandatory": {
            "type": "boolean"
          }
        }
      },
      "LogoutAllDto": {
        "type": "object",
        "properties": {
          "stepUpChallengeId": {
            "type": "string",
            "description": "از POST /auth/step-up/request (scope: SESSION_REVOKE)"
          },
          "stepUpCode": {
            "type": "string",
            "example": "482913"
          }
        },
        "required": [
          "stepUpChallengeId",
          "stepUpCode"
        ]
      },
...
      "CreateBlogPostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "راهنمای کامل چک‌این آنلاین پروازهای داخلی"
          },
          "body": {
            "type": "string",
            "example": "متن مقاله…"
          },
          "category": {
            "type": "string",
            "enum": [
              "NEWS",
              "GUIDE",
              "DEST",
              "OFFERS"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "SCHEDULED"
            ],
            "default": "DRAFT"
          },
          "coverFileId": {
            "type": "string",
            "description": "StoredFile id from POST /files"
          },
          "scheduledAt": {
            "type": "string",
            "description": "Required when status is SCHEDULED (ISO 8601 UTC)"
          },
          "slug": {
            "type": "string",
            "description": "URL slug — auto-generated if omitted"
          }
        },
        "required": [
          "title",
          "body",
          "category"
        ]
      },
      "UpdateBlogPostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "NEWS",
              "GUIDE",
              "DEST",
              "OFFERS"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "SCHEDULED"
            ]
          },
          "coverFileId": {
            "type": "object",
            "nullable": true
          },
          "scheduledAt": {
            "type": "object"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
    }
  }
}